fixed IT tests after introduction of presync
