+++ internal/errors/errors_test.go
@@ -5,51 +5,68 @@
 package errors
 
 import (
+	"errors"
 	"fmt"
 	"testing"
 	"time"
 
+	druidv1alpha1 "github.com/gardener/etcd-druid/api/v1alpha1"
 	. "github.com/onsi/gomega"
 )
 
 func TestWrapError(t *testing.T) {
 	err := WrapError(fmt.Errorf("testError"), "ERR_TEST", "testOp", "testMsg")
-
 	g := NewWithT(t)
-	druidErr, ok := err.(*DruidError)
-	g.Expect(ok).To(BeTrue())
+	druidErr := &DruidError{}
+	g.Expect(errors.As(err, &druidErr)).To(BeTrue())
 	g.Expect(string(druidErr.Code)).To(Equal("ERR_TEST"))
 	g.Expect(druidErr.Operation).To(Equal("testOp"))
 	g.Expect(druidErr.Message).To(Equal("testMsg"))
 }
 
-func TestIsRetriable(t *testing.T) {
-	err := &DruidError{
-		Code:       ErrRetriable,
-		Cause:      nil,
-		Operation:  "",
-		Message:    "",
-		ObservedAt: time.Now().UTC(),
+func TestIsRequeueAfterError(t *testing.T) {
+	testCases := []struct {
+		name     string
+		code     druidv1alpha1.ErrorCode
+		expected bool
+	}{
+		{
+			name:     "error has code ERR_REQUEUE_AFTER",
+			code:     ErrRequeueAfter,
+			expected: true,
+		},
+		{
+			name:     "error has code ERR_TEST",
+			code:     druidv1alpha1.ErrorCode("ERR_TEST"),
+			expected: false,
+		},
 	}
 
 	g := NewWithT(t)
-	g.Expect(IsRetriable(err)).To(BeTrue())
+	t.Parallel()
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			err := &DruidError{
+				Code:       tc.code,
+				ObservedAt: time.Now().UTC(),
+			}
+			g.Expect(IsRequeueAfterError(err)).To(Equal(tc.expected))
 
-	err.Code = "ERR_TEST"
-	g.Expect(IsRetriable(err)).To(BeFalse())
+		})
+	}
 }
 
 func TestMapToLastErrors(t *testing.T) {
 	errs := []error{
 		&DruidError{
-			Code:       "ERR_TEST1",
+			Code:       druidv1alpha1.ErrorCode("ERR_TEST1"),
 			Cause:      fmt.Errorf("testError1"),
 			Operation:  "testOp",
 			Message:    "testMsg",
 			ObservedAt: time.Now().UTC(),
 		},
 		&DruidError{
-			Code:       "ERR_TEST2",
+			Code:       druidv1alpha1.ErrorCode("ERR_TEST2"),
 			Cause:      nil,
 			Operation:  "testOp",
 			Message:    "testMsg",
